+++ rust/pyxel-wrapper/src/music_wrapper.rs
@@ -80,7 +80,7 @@ impl Music {
     }
 
     #[pyo3(signature = (filename, sec, ffmpeg=None))]
-    pub fn save(&self, filename: &str, sec: f64, ffmpeg: Option<bool>) {
+    pub fn save(&self, filename: &str, sec: f32, ffmpeg: Option<bool>) {
         self.inner.lock().save(filename, sec, ffmpeg);
     }
